agents/python-agent/collector/system_metrics.py
import asyncio
import psutil
import json

from ping3 import ping
import ifcfg 
import platform
import pynvml
import pyadl
import subprocess
import psycopg2
import os
import datetime


def get_gpu_stats():
    gpu_info = {}
    try:
        pynvml.nvmlInit()
        device_count = pynvml.nvmlDeviceGetCount()
        gpus = []
        for i in range(device_count):
            handle = pynvml.nvmlDeviceGetHandleByIndex(i)
            name = pynvml.nvmlDeviceGetName(handle).decode()
            memory = pynvml.nvmlDeviceGetMemoryInfo(handle)
            temperature = pynvml.nvmlDeviceGetTemperature(handle, pynvml.NVML_TEMPERATURE_GPU)
            utilization = pynvml.nvmlDeviceGetUtilizationRates(handle)
            gpus.append({
                'name': name,
                'memory_total': memory.total,
                'memory_used': memory.used,
                'memory_free': memory.free,
                'temperature': temperature,
                'gpu_utilization': utilization.gpu,
                'memory_utilization': utilization.memory
            })
        gpu_info['vendor'] = 'NVIDIA'
        gpu_info['gpus'] = gpus 
        pynvml.nvmlShutdown()
    except Exception as e:
        gpu_info['error'] = str(e)
    return gpu_info

def get_ping():
    try:
        interfaces = ifcfg.interfaces()
        ip = None
        for iface in interfaces.values():
            if 'inet' in iface and iface['inet'] != '127.0.0.1':
                ip = iface['inet']
                break
        if ip:
            result = ping(ip)
            return result * 1000000 if result else None
        return None
    except Exception as e:
        print("error: {e}")
        return None

def gather_cpu_times():
    cpu_times = psutil.cpu_times(percpu=True)

    user_time = {f"core_{i+1}": core[0] for i, core in enumerate(cpu_times)}
    system_time = {f"core_{i+1}": core[1] for i, core in enumerate(cpu_times)}
    idle_time = {f"core_{i+1}": core[2] for i, core in enumerate(cpu_times)}
    
    return user_time, system_time, idle_time

def gather_cpu_percents():
    cpu_percents = psutil.cpu_percent(percpu=True)
    return {f"core_{i+1}": percent for i, percent in enumerate(cpu_percents)}

def gather_virtual_memory_stats():
    virtual_memory = psutil.virtual_memory()
    return virtual_memory[1], virtual_memory[2], virtual_memory[3]

def gather_swap_memory_stats():
    swap_memory = psutil.swap_memory()
    return swap_memory[1], swap_memory[2], swap_memory[3]

def get_disk_usage():
    disk_partitions = psutil.disk_partitions()
    disk_usage_info = {}
    for partition in disk_partitions:
        try:
            usage = psutil.disk_usage(partition.mountpoint)
            disk_usage_info[partition.device] = {
                "mountpoint": partition.mountpoint,
                "fstype": partition.fstype,
                "total": usage.total,
                "used": usage.used,
                "free": usage.free,
                "percent": usage.percent
            }
        except PermissionError:
            continue
    return disk_usage_info

# Windows needs diskperf -y ran in cmd.exe
def get_disk_io_counters():
    raw_disks = psutil.disk_io_counters(perdisk=True)
    result = {}
    for key, value in raw_disks.items():
        result[key] = {
            'read_count': value.read_count,
            'write_count': value.write_count,
            'read_bytes': value.read_bytes,
            'write_bytes': value.write_bytes,
            'read_time': value.read_time,
            'write_time': value.write_time,
        }
    return result

def collect_metrics():

    system_info = {}

    # CPU
    system_info['cpu'] = {}
    user_time, system_time, idle_time = gather_cpu_times()
    system_info['cpu']['user_time'] = user_time
    system_info['cpu']['system_time'] = system_time
    system_info['cpu']['idle_time'] = idle_time
    system_info['cpu']['percent'] = gather_cpu_percents()

    # Memory
    system_info['memory'] = {}
    available_memory, percent_usage, used_memory = gather_virtual_memory_stats()
    system_info['memory']['available_memory'] = available_memory
    system_info['memory']['memory_percent_usage'] = percent_usage
    system_info['memory']['used_memory'] = used_memory

    # Swap memory
    system_info['swap_memory'] = {}
    swap_used_memory, swap_free_memory, swap_percent_usage = gather_swap_memory_stats()
    system_info['swap_memory']['used_memory'] = swap_used_memory
    system_info['swap_memory']['free_memory'] = swap_free_memory
    system_info['swap_memory']['percent_usage'] = swap_percent_usage

    # Disk usage
    system_info['disk_usage'] = get_disk_usage()

    # IO
    system_info['io'] = get_disk_io_counters()

    return system_info

async def async_collect_metrics():
    loop = asyncio.get_running_loop()
    
    # Run the blocking function in a separate thread
    metrics = await loop.run_in_executor(None, collect_metrics)
    return metrics