agents/python-agent/requirements.txt
grpcio
grpcio-tools
protobuf
psutil
ping3
ifcfg
pynvml
pyadl
psycopg2-binary